+++ packages/expo-modules-core/CHANGELOG.md
@@ -8,6 +8,8 @@
 
 ### 🐛 Bug fixes
 
+- [iOS] Fixed URI with encoded `#` getting cut off.
+
 ### 💡 Others
 
 ## 1.2.2 — 2023-02-14